git-svn-id: https://yuma.svn.sourceforge.net/svnroot/yuma/trunk@1441 1ffd1d58-2efc-45df-af60-190a33ec9863

+++ netconf/src/ncx/ncx_num.c
@@ -2209,7 +2209,7 @@ status_t
             }
         }
 #else
-        ilen = sprintf((char *)buff, "%lld", num->d);
+        ilen = sprintf((char *)buff, "%lld", (long long)num->d);
 #endif
         break;
     default:
